Sépare Chart et ChartModal en 2 fichiers différents
